pyproject.toml
[project]
name = "datacontract-cli"
version = "0.10.7"
description = "Test data contracts"
readme = "README.md"
authors = [
  { name = "Jochen Christ", email = "jochen.christ@innoq.com" },
  { name = "Stefan Negele", email = "stefan.negele@innoq.com" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.9,<0.13",
  "pydantic>=2.5.3,<2.8.0",
  "pyyaml~=6.0.1",
  "requests>=2.31,<2.33",
  "fastapi==0.111.0",
  "fastparquet==2024.5.0",
  "python-multipart==0.0.9",
  "rich~=13.7.0",
  "simple-ddl-parser==1.5.1",
  "soda-core-duckdb>=3.3.1,<3.4.0",
  # remove setuptools when https://github.com/sodadata/soda-core/issues/2091 is resolved
  "setuptools>=60",
  "duckdb==1.0.0",
  "fastjsonschema~=2.19.1",
  "python-dotenv~=1.0.0",
  "rdflib==7.0.0",
  "opentelemetry-exporter-otlp-proto-grpc~=1.16",
  "opentelemetry-exporter-otlp-proto-http~=1.16",
  "boto3>=1.34.41,<1.34.120",
  "botocore>=1.34.41,<1.34.120",
  "jinja_partials >= 0.2.1"
]

[project.optional-dependencies]

avro = [
  "avro==1.11.3"
]

bigquery = [
  "soda-core-bigquery>=3.3.1,<3.4.0"
]



databricks = [
  "soda-core-spark-df>=3.3.1,<3.4.0",
  "databricks-sql-connector>=3.1.2,<3.2.0",
  "soda-core-spark[databricks]>=3.3.1,<3.4.0"
]

deltalake = [
    "deltalake>=0.17,<0.19"
]

kafka = [
  "datacontract-cli[avro]",
  "soda-core-spark-df>=3.3.1,<3.4.0"
]

postgres = [
  "soda-core-postgres>=3.3.1,<3.4.0"
]

s3 = [
  "s3fs==2024.6.0"
]

snowflake = [
  "snowflake-connector-python[pandas]>=3.6,<3.11",
  "soda-core-snowflake>=3.3.1,<3.4.0"
]

sqlserver = [
    "soda-core-sqlserver>=3.3.1,<3.4.0"
]

trino = [
  "soda-core-trino>=3.3.1,<3.4.0"
]

all = [
  "datacontract-cli[kafka,bigquery,snowflake,postgres,databricks,sqlserver,s3,deltalake,trino]"
]

dev = [
  "datacontract-cli[all]",
  "httpx==0.27.0",
  "ruff",
  "pre-commit~=3.7.1",
  "pytest",
  "pytest-xdist",
  "moto",
  "pymssql==2.3.0",
  "kafka-python",
  "trino==0.328.0",
  "testcontainers~=4.5.0",
  "testcontainers[core]",
  "testcontainers[minio]",
  "testcontainers[postgres]",
  "testcontainers[kafka]",
  "testcontainers[mssql]",
]

[project.urls]
Homepage = "https://cli.datacontract.com"
Issues = "https://github.com/datacontract/cli/issues"

[project.scripts]
datacontract = "datacontract.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# [tool.pytest.ini_options]
# addopts="-n 8"

[tool.ruff]
line-length = 120
